commit new

+++ app/Http/Controllers/CustomerController.php
@@ -2,11 +2,16 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\Controller;
+use Illuminate\Foundation\Auth\RegistersUsers;
 use Illuminate\Http\Request;
 use App;
 use App\Customer;
+use App\User;
 use DateTime;
+use Validator;
 use Auth;
+use App\Http\Middleware\CheckAdmin;
 class CustomerController extends Controller
 {
     /**
@@ -17,9 +22,13 @@ class CustomerController extends Controller
      public function __construct()
      {
          $this->middleware('auth');
+
      }
     public function index()
     {
+      if(Auth::user()->role!='admin'){
+        return redirect('/');
+      }
       $customers = Customer::all();
 
       return view('customer.index')->with('values', $customers);
@@ -32,6 +41,9 @@ class CustomerController extends Controller
      */
     public function create()
     {
+      if(Auth::user()->role!='admin'){
+        return redirect('/');
+      }
          return view('customer.create');
     }
 
@@ -43,6 +55,9 @@ class CustomerController extends Controller
      */
     public function store(Request $request)
     {
+      if(Auth::user()->role!='admin'){
+        return redirect('/');
+      }
         $this->validate($request, [
              'firstname' => 'required|max:255',
              'lastname' => 'required',
@@ -52,21 +67,23 @@ class CustomerController extends Controller
          if( $request->password!==$request->confirmpassword){
            return view('customer.create')->withInput();
          }
+
+         $user = new User;
+         $user->username = $request->username;
+         $user->password = bcrypt($request->password);
+         $user->email = "-";
+         $user->name = $request->firstname.' '.$request->lastname;
+         $user->role =$request->role;
+         $user->save();
         $customer = new Customer;
         $now = new DateTime();
         $customer->firstname = $request->firstname;
         $customer->lastname = $request->lastname;
         $customer->companyname = $request->companyname;
-        $customer->username = $request->username;
-        // $customer->address = $request->address;
-        // $customer->tel = $request->tel;
-        // $customer->fax = $request->fax;
-
-        $customer->password = $request->password;
-        // $customer->taxno = $request->taxno;
+        $customer->aboutwork = $request->aboutwork;
+        $customer->user_id = $user->id;
         $customer->date =$now;
         $customer->by = Auth::user()->name;
-
         $customer->save();
         return redirect()->route('customer.index');
     }
@@ -79,6 +96,10 @@ class CustomerController extends Controller
      */
     public function show($id)
     {
+
+      if(Auth::user()->role!='admin'){
+        return redirect('/');
+      }
       $customer = Customer::find($id);
 
        return view('customer.show')
@@ -93,6 +114,9 @@ class CustomerController extends Controller
      */
     public function edit($id)
     {
+      if(Auth::user()->role!='admin'){
+        return redirect('/');
+      }
       $customer = Customer::find($id);
 
        return view('customer.edit')
@@ -108,6 +132,19 @@ class CustomerController extends Controller
      */
     public function update(Request $request, $id)
     {
+      $now = new DateTime();
+      $validator = Validator::make($request->all(), [
+            'password' => 'required|min:6|confirmed',
+            'taxno' => 'required',
+            'tel' => 'required'
+        ]);
+
+        if ($validator->fails()) {
+                return redirect('/addmoreinfo')
+                            ->withErrors($validator)
+                            ->withInput();
+            }
+
       $customer = App\Customer::find($id);
       $customer->firstname = $request->firstname;
       $customer->lastname = $request->lastname;
@@ -115,12 +152,15 @@ class CustomerController extends Controller
       $customer->address = $request->address;
       $customer->tel = $request->tel;
       $customer->fax = $request->fax;
-      $customer->email = $request->email;
-      $customer->password = $request->password;
+
       $customer->taxno = $request->taxno;
-      $customer->date = $request->date;
+      $customer->date = $now;
       $customer->save();
-      return redirect()->route('customer.index');
+
+      $user = App\User::find($customer->user_id);
+      $user->password = bcrypt($request->password);
+      $user->save();
+      return redirect('/');
 }
     /**
      * Remove the specified resource from storage.
@@ -130,6 +170,9 @@ class CustomerController extends Controller
      */
     public function destroy($id)
     {
+      if(Auth::user()->role!='admin'){
+        return redirect('/');
+      }
       $customer = Customer::find($id);
       $customer->delete();
       return redirect()->route('customer.index');
